requirements.txt
# Core testing
pytest==9.0.3
pytest-html==4.2.0
pytest-metadata==3.1.1

# Allure reporting
allure-pytest==2.15.3
allure-python-commons==2.15.3

# Parallel execution
pytest-xdist==3.8.0

# Test reliability
pytest-rerunfailures==16.1
pytest-timeout==2.3.1

# Logging
loguru==0.7.3
colorama==0.4.6

# Configuration
python-dotenv==1.2.2
PyYAML==6.0.3

# HTTP client
requests==2.33.1

# HTML templating
Jinja2==3.1.6
MarkupSafe==3.0.3

# Utilities
packaging==26.1
pluggy==1.6.0

# Test dependencies (pin versions for reproducibility)
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
execnet==2.1.2
idna==3.13
iniconfig==2.3.0
Pygments==2.20.0
urllib3==2.6.3

# Windows-specific
win32_setctime==1.2.0
